requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
black==24.10.0
blinker==1.9.0
cachelib==0.9.0
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
Mako==1.3.7
MarkupSafe==3.0.2
mypy-extensions==1.0.0
numpy==2.0.2
openai==1.57.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
scikit-learn==1.6.0
scipy==1.13.1
sniffio==1.3.1
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3<2.0
Werkzeug==3.1.3
zipp==3.21.0
